secondd commit

+++ contracts/IUniswap.sol
@@ -3,54 +3,68 @@ pragma solidity ^0.8.9;
 
 interface IUniswap {
     function swapTokensForExactTokens(
-        uint amountOut,
-        uint amountInMax,
+        uint256 amountOut,
+        uint256 amountInMax,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
-    function swapExactETHForTokens(uint amountOutMin, address[] calldata path, address to, uint deadline)
-  external
-  payable
-  returns (uint[] memory amounts);
+    function swapExactETHForTokens(
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external payable returns (uint256[] memory amounts);
 
-
-  function swapTokensForExactETH(uint amountOut, uint amountInMax, address[] calldata path, address to, uint deadline)
-  external
-  returns (uint[] memory amounts);
-
+    function swapTokensForExactETH(
+        uint256 amountOut,
+        uint256 amountInMax,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
     function addLiquidity(
-  address tokenA,
-  address tokenB,
-  uint amountADesired,
-  uint amountBDesired,
-  uint amountAMin,
-  uint amountBMin,
-  address to,
-  uint deadline
-) external returns (uint amountA, uint amountB, uint liquidity);
+        address tokenA,
+        address tokenB,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
+        address to,
+        uint256 deadline
+    )
+        external
+        returns (
+            uint256 amountA,
+            uint256 amountB,
+            uint256 liquidity
+        );
 
-function addLiquidityETH(
-  address token,
-  uint amountTokenDesired,
-  uint amountTokenMin,
-  uint amountETHMin,
-  address to,
-  uint deadline
-) external payable returns (uint amountToken, uint amountETH, uint liquidity);
+    function addLiquidityETH(
+        address token,
+        uint256 amountTokenDesired,
+        uint256 amountTokenMin,
+        uint256 amountETHMin,
+        address to,
+        uint256 deadline
+    )
+        external
+        payable
+        returns (
+            uint256 amountToken,
+            uint256 amountETH,
+            uint256 liquidity
+        );
 
-function removeLiquidity(
-  address tokenA,
-  address tokenB,
-  uint liquidity,
-  uint amountAMin,
-  uint amountBMin,
-  address to,
-  uint deadline
-) external returns (uint amountA, uint amountB);
-
-
-   
+    function removeLiquidity(
+        address tokenA,
+        address tokenB,
+        uint256 liquidity,
+        uint256 amountAMin,
+        uint256 amountBMin,
+        address to,
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB);
 }
